+++ src/app/App.js
@@ -1,0 +1,47 @@
+import { Component } from 'react';
+import Header from './Components/Header';
+import Footer from './Components/Footer';
+import Home from '../home/Home';
+import {
+  BrowserRouter as Router,
+  Route,
+  Switch,
+  Redirect
+} from 'react-router-dom';
+import './App.css';
+import Login from './Login';
+
+
+class App extends Component {
+
+  render() {
+    return (
+      <div className="App">
+        <Router>
+          <Header/>
+          <main>
+
+            <Switch>
+              <Route path="/" exact={true}
+                render={routerProps => (
+                  <Home {...routerProps}/>
+                )}
+              />
+              <Route path="/login" exact={true}
+                render={routerProps => (
+                  <Login {...routerProps}/>
+                )}
+              />
+              <Redirect to="/" />
+
+            </Switch>
+          </main>
+          <Footer/>
+        </Router>
+      </div>
+    );
+  }
+
+}
+
+export default App;
