+++ src/app/Components/Footer.css
@@ -1,0 +1,5 @@
+.Footer {
+  padding: 5px;
+  font-size: .8rem;
+  background: rgba(0, 0, 0, 0.1);
+}
