+++ src/app/Components/Footer.js
@@ -1,0 +1,15 @@
+import { Component } from 'react';
+import './Footer.css';
+
+class Footer extends Component {
+  
+  render() { 
+    return (
+      <footer className="Footer">
+        &copy; 2021 You
+      </footer>
+    );
+  }
+}
+ 
+export default Footer;
